+++ packages/messenger-internal/src/index.ts
@@ -1,5 +1,5 @@
-export * from "./ActionResult";
-export * from "./v1/Observable";
+export * from "./shared/Result";
+export * from "./shared/Observable";
 export * from "./types";
 export * from "./versioning";
 export * from "./loader";
@@ -7,3 +7,4 @@ export * from "./legacy-loader";
 
 export * as v0 from "./v0";
 export * as v1 from "./v1";
+export * as v2 from "./v2";