+++ packages/messenger-internal/src/loader/types.ts
@@ -1,8 +1,9 @@
-import type { ActionResult } from "../ActionResult";
 import type { MessengerInfo } from "../types";
 import type { VersionedApi } from "../versioning";
 import type * as v0 from "../v0";
 import type * as v1 from "../v1";
+import type * as v2 from "../v2";
+import { Result } from "../shared";
 
 export interface WidgetLoaderSettings {
   widget_key: string;
@@ -11,24 +12,28 @@ export interface WidgetLoaderSettings {
 export interface WidgetLoader {
   load: (opts?: LegacyOptions) => Promise<void>;
   createMessenger(
-    version: 0
+    version: 0,
   ): (
-    settings: WidgetLoaderSettings
-  ) => Promise<ActionResult<string, MessengerInfo<0, v0.Api>>>;
+    settings: WidgetLoaderSettings,
+  ) => Promise<Result<MessengerInfo<0, v0.Api>, string>>;
 
   createMessenger(
-    version: 1
+    version: 1,
   ): (
-    settings: WidgetLoaderSettings
-  ) => Promise<ActionResult<string, MessengerInfo<1, v1.Api>>>;
+    settings: WidgetLoaderSettings,
+  ) => Promise<Result<MessengerInfo<1, v1.Api>, string>>;
 
   createMessenger(
-    version: number
+    version: 2,
   ): (
-    settings: WidgetLoaderSettings
-  ) => Promise<
-    ActionResult<string, MessengerInfo<number, VersionedApi<number>>>
-  >;
+    settings: WidgetLoaderSettings,
+  ) => Promise<Result<MessengerInfo<2, v2.Api>, string>>;
+
+  createMessenger(
+    version: number,
+  ): (
+    settings: WidgetLoaderSettings,
+  ) => Promise<Result<MessengerInfo<number, VersionedApi<number>>, string>>;
 }
 
 export interface LegacyOptions {
@@ -38,5 +43,5 @@ export interface LegacyOptions {
 
 export type Run = (
   settings: WidgetLoaderSettings,
-  hostWindow: Window
+  hostWindow: Window,
 ) => Promise<WidgetLoader>;